moved span receiver from example to library
